[gl] Refactor c++ native layer
