+++ packages/expo-gl/android/src/main/cpp/EXGLJniApi.cpp
@@ -1,10 +1,6 @@
-#include <stdint.h>
-
 #include <jni.h>
 #include <thread>
-#include <android/log.h>
 
-#include <jsi/jsi.h>
 #include "EXGLNativeApi.h"
 #include "EXPlatformUtils.h"
 
@@ -12,7 +8,7 @@ extern "C" {
 
 // JNIEnv is valid only inside the same thread that it was passed from
 // to support worklet we need register it from UI thread
-thread_local JNIEnv* threadLocalEnv;
+thread_local JNIEnv *threadLocalEnv;
 
 JNIEXPORT jint JNICALL
 Java_expo_modules_gl_cpp_EXGL_EXGLContextCreate
@@ -31,7 +27,7 @@ Java_expo_modules_gl_cpp_EXGL_EXGLContextPrepare
   std::function<void(void)> flushMethod = [glContextRef, flushMethodRef] {
     threadLocalEnv->CallVoidMethod(glContextRef, flushMethodRef);
   };
-  EXGLContextPrepare((void*) jsiPtr, exglCtxId, flushMethod);
+  EXGLContextPrepare((void *) jsiPtr, exglCtxId, flushMethod);
 }
 
 JNIEXPORT void JNICALL
@@ -77,7 +73,7 @@ Java_expo_modules_gl_cpp_EXGL_EXGLContextGetObject
   return EXGLContextGetObject(exglCtxId, exglObjId);
 }
 
-JNIEXPORT bool JNICALL
+JNIEXPORT jboolean JNICALL
 Java_expo_modules_gl_cpp_EXGL_EXGLContextNeedsRedraw
 (JNIEnv *env, jclass clazz, jint exglCtxId) {
   return EXGLContextNeedsRedraw(exglCtxId);
