+++ packages/expo-gl/common/EXGLImageUtils.h
@@ -1,15 +1,6 @@
 #pragma once
 
-#ifdef __ANDROID__
-#include <GLES3/gl3.h>
-#include <GLES3/gl3ext.h>
-#endif
-#ifdef __APPLE__
-#include <OpenGLES/ES3/gl.h>
-#endif
-
-#include <jsi/jsi.h>
-#include <vector>
+#include "pch.h"
 
 namespace expo {
 namespace gl_cpp {
@@ -19,10 +10,10 @@ GLuint bytesPerPixel(GLenum type, GLenum format);
 void flipPixels(GLubyte *pixels, size_t bytesPerRow, size_t rows);
 
 std::shared_ptr<uint8_t> loadImage(
-    facebook::jsi::Runtime &runtime,
-    const facebook::jsi::Object &jsPixels,
-    int *fileWidth,
-    int *fileHeight,
-    int *fileComp);
+  facebook::jsi::Runtime &runtime,
+  const facebook::jsi::Object &jsPixels,
+  int *fileWidth,
+  int *fileHeight,
+  int *fileComp);
 } // namespace gl_cpp
 } // namespace expo
