+++ packages/expo-gl/common/EXGLNativeApi.cpp
@@ -9,9 +9,9 @@ EXGLContextId EXGLContextCreate() {
 }
 
 void EXGLContextPrepare(
-    void *jsiPtr,
-    EXGLContextId exglCtxId,
-    std::function<void(void)> flushMethod) {
+  void *jsiPtr,
+  EXGLContextId exglCtxId,
+  std::function<void(void)> flushMethod) {
   auto [exglCtx, lock] = ContextGet(exglCtxId);
   if (exglCtx) {
     exglCtx->prepareContext(*reinterpret_cast<jsi::Runtime *>(jsiPtr), flushMethod);
