+++ packages/expo-gl/common/EXJsiUtils.h
@@ -1,7 +1,6 @@
 #pragma once
 
-#include <jsi/jsi.h>
-#include <vector>
+#include "pch.h"
 
 #include "EXPlatformUtils.h"
 #include "EXTypedArrayApi.h"
@@ -10,7 +9,7 @@ namespace expo {
 namespace gl_cpp {
 namespace jsi = facebook::jsi;
 
-template <typename T>
+template<typename T>
 inline std::vector<T> jsArrayToVector(jsi::Runtime &runtime, const jsi::Array &jsArray) {
   size_t length = jsArray.length(runtime);
   std::vector<T> values(length);
@@ -21,7 +20,7 @@ inline std::vector<T> jsArrayToVector(jsi::Runtime &runtime, const jsi::Array &j
   return values;
 }
 
-template <>
+template<>
 inline std::vector<std::string> jsArrayToVector(jsi::Runtime &runtime, const jsi::Array &jsArray) {
   size_t length = jsArray.length(runtime);
   std::vector<std::string> strings(length);
@@ -44,37 +43,39 @@ inline std::vector<uint8_t> rawTypedArray(jsi::Runtime &runtime, const jsi::Obje
 
 inline bool jsValueToBool(jsi::Runtime &runtime, const jsi::Value &jsValue) {
   return jsValue.isBool()
-      ? jsValue.getBool()
-      : throw std::runtime_error(jsValue.toString(runtime).utf8(runtime) + " is not a bool value");
+         ? jsValue.getBool()
+         : throw std::runtime_error(
+      jsValue.toString(runtime).utf8(runtime) + " is not a bool value"
+    );
 }
 
-template <typename Func>
+template<typename Func>
 inline void setFunctionOnObject(
-    jsi::Runtime &runtime,
-    jsi::Object &jsObject,
-    const std::string &name,
-    Func func) {
+  jsi::Runtime &runtime,
+  jsi::Object &jsObject,
+  const std::string &name,
+  Func func) {
   auto jsName = jsi::PropNameID::forUtf8(runtime, name);
   jsObject.setProperty(
-      runtime, jsName, jsi::Function::createFromHostFunction(runtime, jsName, 0, func));
+    runtime, jsName, jsi::Function::createFromHostFunction(runtime, jsName, 0, func));
 }
 
 inline void jsConsoleLog(jsi::Runtime &runtime, std::initializer_list<jsi::Value> args) {
   runtime.global()
-      .getProperty(runtime, "console")
-      .asObject(runtime)
-      .getProperty(runtime, "log")
-      .asObject(runtime)
-      .asFunction(runtime)
-      .call(runtime, args);
+    .getProperty(runtime, "console")
+    .asObject(runtime)
+    .getProperty(runtime, "log")
+    .asObject(runtime)
+    .asFunction(runtime)
+    .call(runtime, args);
 }
 
 inline jsi::Value unsupportedWebGL2(
-    const std::string &name,
-    jsi::Runtime &runtime,
-    const jsi::Value &jsThis,
-    const jsi::Value *jsArgv,
-    size_t argc) {
+  const std::string &name,
+  jsi::Runtime &runtime,
+  const jsi::Value &jsThis,
+  const jsi::Value *jsArgv,
+  size_t argc) {
   throw std::runtime_error("EXGL: This device doesn't support WebGL2 method: " + name + "()!");
 }
