+++ packages/expo-gl/common/EXTypedArrayApi.cpp
@@ -1,11 +1,9 @@
 #include "EXTypedArrayApi.h"
 
-#include <unordered_map>
-
 namespace expo {
 namespace gl_cpp {
 
-template <TypedArrayKind T>
+template<TypedArrayKind T>
 using ContentType = typename typedArrayTypeMap<T>::type;
 
 enum class Prop {
@@ -30,7 +28,7 @@ enum class Prop {
 };
 
 class PropNameIDCache {
- public:
+public:
   const jsi::PropNameID &get(jsi::Runtime &runtime, Prop prop) {
     auto key = reinterpret_cast<uintptr_t>(&runtime);
     if (this->props.find(key) == this->props.end()) {
@@ -50,7 +48,7 @@ class PropNameIDCache {
     }
   }
 
- private:
+private:
   std::unordered_map<uintptr_t, std::unordered_map<Prop, std::unique_ptr<jsi::PropNameID>>> props;
 
   jsi::PropNameID createProp(jsi::Runtime &runtime, Prop prop);
@@ -61,6 +59,7 @@ PropNameIDCache propNameIDCache;
 InvalidateCacheOnDestroy::InvalidateCacheOnDestroy(jsi::Runtime &runtime) {
   key = reinterpret_cast<uintptr_t>(&runtime);
 }
+
 InvalidateCacheOnDestroy::~InvalidateCacheOnDestroy() {
   propNameIDCache.invalidate(key);
 }
@@ -68,24 +67,24 @@ InvalidateCacheOnDestroy::~InvalidateCacheOnDestroy() {
 TypedArrayKind getTypedArrayKindForName(const std::string &name);
 
 TypedArrayBase::TypedArrayBase(jsi::Runtime &runtime, size_t size, TypedArrayKind kind)
-    : TypedArrayBase(
-          runtime,
-          runtime.global()
-              .getProperty(runtime, propNameIDCache.getConstructorNameProp(runtime, kind))
-              .asObject(runtime)
-              .asFunction(runtime)
-              .callAsConstructor(runtime, {static_cast<double>(size)})
-              .asObject(runtime)) {}
+  : TypedArrayBase(
+  runtime,
+  runtime.global()
+    .getProperty(runtime, propNameIDCache.getConstructorNameProp(runtime, kind))
+    .asObject(runtime)
+    .asFunction(runtime)
+    .callAsConstructor(runtime, {static_cast<double>(size)})
+    .asObject(runtime)) {}
 
 TypedArrayBase::TypedArrayBase(jsi::Runtime &runtime, const jsi::Object &obj)
-    : jsi::Object(jsi::Value(runtime, obj).asObject(runtime)) {}
+  : jsi::Object(jsi::Value(runtime, obj).asObject(runtime)) {}
 
 TypedArrayKind TypedArrayBase::getKind(jsi::Runtime &runtime) const {
   auto constructorName = this->getProperty(runtime, propNameIDCache.get(runtime, Prop::Constructor))
-                             .asObject(runtime)
-                             .getProperty(runtime, propNameIDCache.get(runtime, Prop::Name))
-                             .asString(runtime)
-                             .utf8(runtime);
+    .asObject(runtime)
+    .getProperty(runtime, propNameIDCache.get(runtime, Prop::Name))
+    .asString(runtime)
+    .utf8(runtime);
   return getTypedArrayKindForName(constructorName);
 };
 
@@ -127,12 +126,12 @@ jsi::ArrayBuffer TypedArrayBase::getBuffer(jsi::Runtime &runtime) const {
 
 bool isTypedArray(jsi::Runtime &runtime, const jsi::Object &jsObj) {
   auto jsVal = runtime.global()
-                   .getProperty(runtime, propNameIDCache.get(runtime, Prop::ArrayBuffer))
-                   .asObject(runtime)
-                   .getProperty(runtime, propNameIDCache.get(runtime, Prop::IsView))
-                   .asObject(runtime)
-                   .asFunction(runtime)
-                   .callWithThis(runtime, runtime.global(), {jsi::Value(runtime, jsObj)});
+    .getProperty(runtime, propNameIDCache.get(runtime, Prop::ArrayBuffer))
+    .asObject(runtime)
+    .getProperty(runtime, propNameIDCache.get(runtime, Prop::IsView))
+    .asObject(runtime)
+    .asFunction(runtime)
+    .callWithThis(runtime, runtime.global(), {jsi::Value(runtime, jsObj)});
   if (jsVal.isBool()) {
     return jsVal.getBool();
   } else {
@@ -142,12 +141,12 @@ bool isTypedArray(jsi::Runtime &runtime, const jsi::Object &jsObj) {
 
 TypedArrayBase getTypedArray(jsi::Runtime &runtime, const jsi::Object &jsObj) {
   auto jsVal = runtime.global()
-                   .getProperty(runtime, propNameIDCache.get(runtime, Prop::ArrayBuffer))
-                   .asObject(runtime)
-                   .getProperty(runtime, propNameIDCache.get(runtime, Prop::IsView))
-                   .asObject(runtime)
-                   .asFunction(runtime)
-                   .callWithThis(runtime, runtime.global(), {jsi::Value(runtime, jsObj)});
+    .getProperty(runtime, propNameIDCache.get(runtime, Prop::ArrayBuffer))
+    .asObject(runtime)
+    .getProperty(runtime, propNameIDCache.get(runtime, Prop::IsView))
+    .asObject(runtime)
+    .asFunction(runtime)
+    .callWithThis(runtime, runtime.global(), {jsi::Value(runtime, jsObj)});
   if (jsVal.isBool()) {
     return TypedArrayBase(runtime, jsObj);
   } else {
@@ -163,15 +162,15 @@ std::vector<uint8_t> arrayBufferToVector(jsi::Runtime &runtime, jsi::Object &jsO
 
   uint8_t *dataBlock = jsArrayBuffer.data(runtime);
   size_t blockSize =
-      jsArrayBuffer.getProperty(runtime, propNameIDCache.get(runtime, Prop::ByteLength)).asNumber();
+    jsArrayBuffer.getProperty(runtime, propNameIDCache.get(runtime, Prop::ByteLength)).asNumber();
   return std::vector<uint8_t>(dataBlock, dataBlock + blockSize);
 }
 
 void arrayBufferUpdate(
-    jsi::Runtime &runtime,
-    jsi::ArrayBuffer &buffer,
-    std::vector<uint8_t> data,
-    size_t offset) {
+  jsi::Runtime &runtime,
+  jsi::ArrayBuffer &buffer,
+  std::vector<uint8_t> data,
+  size_t offset) {
   uint8_t *dataBlock = buffer.data(runtime);
   size_t blockSize = buffer.size(runtime);
   if (data.size() > blockSize) {
@@ -180,27 +179,27 @@ void arrayBufferUpdate(
   std::copy(data.begin(), data.end(), dataBlock + offset);
 }
 
-template <TypedArrayKind T>
-TypedArray<T>::TypedArray(jsi::Runtime &runtime, size_t size) : TypedArrayBase(runtime, size, T){};
+template<TypedArrayKind T>
+TypedArray<T>::TypedArray(jsi::Runtime &runtime, size_t size) : TypedArrayBase(runtime, size, T) {};
 
-template <TypedArrayKind T>
+template<TypedArrayKind T>
 TypedArray<T>::TypedArray(jsi::Runtime &runtime, std::vector<ContentType<T>> data)
-    : TypedArrayBase(runtime, data.size(), T) {
+  : TypedArrayBase(runtime, data.size(), T) {
   update(runtime, data);
 };
 
-template <TypedArrayKind T>
+template<TypedArrayKind T>
 TypedArray<T>::TypedArray(TypedArrayBase &&base) : TypedArrayBase(std::move(base)) {}
 
-template <TypedArrayKind T>
+template<TypedArrayKind T>
 std::vector<ContentType<T>> TypedArray<T>::toVector(jsi::Runtime &runtime) {
   auto start =
-      reinterpret_cast<ContentType<T> *>(getBuffer(runtime).data(runtime) + byteOffset(runtime));
+    reinterpret_cast<ContentType<T> *>(getBuffer(runtime).data(runtime) + byteOffset(runtime));
   auto end = start + size(runtime);
   return std::vector<ContentType<T>>(start, end);
 }
 
-template <TypedArrayKind T>
+template<TypedArrayKind T>
 void TypedArray<T>::update(jsi::Runtime &runtime, const std::vector<ContentType<T>> &data) {
   if (data.size() != size(runtime)) {
     throw jsi::JSError(runtime, "TypedArray can only be updated with a vector of the same size");
@@ -210,8 +209,8 @@ void TypedArray<T>::update(jsi::Runtime &runtime, const std::vector<ContentType<
 }
 
 const jsi::PropNameID &PropNameIDCache::getConstructorNameProp(
-    jsi::Runtime &runtime,
-    TypedArrayKind kind) {
+  jsi::Runtime &runtime,
+  TypedArrayKind kind) {
   switch (kind) {
     case TypedArrayKind::Int8Array:
       return get(runtime, Prop::Int8Array);
@@ -279,30 +278,47 @@ jsi::PropNameID PropNameIDCache::createProp(jsi::Runtime &runtime, Prop prop) {
 }
 
 std::unordered_map<std::string, TypedArrayKind> nameToKindMap = {
-    {"Int8Array", TypedArrayKind::Int8Array},
-    {"Int16Array", TypedArrayKind::Int16Array},
-    {"Int32Array", TypedArrayKind::Int32Array},
-    {"Uint8Array", TypedArrayKind::Uint8Array},
-    {"Uint8ClampedArray", TypedArrayKind::Uint8ClampedArray},
-    {"Uint16Array", TypedArrayKind::Uint16Array},
-    {"Uint32Array", TypedArrayKind::Uint32Array},
-    {"Float32Array", TypedArrayKind::Float32Array},
-    {"Float64Array", TypedArrayKind::Float64Array},
+  {"Int8Array",         TypedArrayKind::Int8Array},
+  {"Int16Array",        TypedArrayKind::Int16Array},
+  {"Int32Array",        TypedArrayKind::Int32Array},
+  {"Uint8Array",        TypedArrayKind::Uint8Array},
+  {"Uint8ClampedArray", TypedArrayKind::Uint8ClampedArray},
+  {"Uint16Array",       TypedArrayKind::Uint16Array},
+  {"Uint32Array",       TypedArrayKind::Uint32Array},
+  {"Float32Array",      TypedArrayKind::Float32Array},
+  {"Float64Array",      TypedArrayKind::Float64Array},
 };
 
 TypedArrayKind getTypedArrayKindForName(const std::string &name) {
   return nameToKindMap.at(name);
 }
 
-template class TypedArray<TypedArrayKind::Int8Array>;
-template class TypedArray<TypedArrayKind::Int16Array>;
-template class TypedArray<TypedArrayKind::Int32Array>;
-template class TypedArray<TypedArrayKind::Uint8Array>;
-template class TypedArray<TypedArrayKind::Uint8ClampedArray>;
-template class TypedArray<TypedArrayKind::Uint16Array>;
-template class TypedArray<TypedArrayKind::Uint32Array>;
-template class TypedArray<TypedArrayKind::Float32Array>;
-template class TypedArray<TypedArrayKind::Float64Array>;
+template
+class TypedArray<TypedArrayKind::Int8Array>;
+
+template
+class TypedArray<TypedArrayKind::Int16Array>;
+
+template
+class TypedArray<TypedArrayKind::Int32Array>;
+
+template
+class TypedArray<TypedArrayKind::Uint8Array>;
+
+template
+class TypedArray<TypedArrayKind::Uint8ClampedArray>;
+
+template
+class TypedArray<TypedArrayKind::Uint16Array>;
+
+template
+class TypedArray<TypedArrayKind::Uint32Array>;
+
+template
+class TypedArray<TypedArrayKind::Float32Array>;
+
+template
+class TypedArray<TypedArrayKind::Float64Array>;
 
 } // namespace gl_cpp
 } // namespace expo
