+++ packages/expo-gl/common/EXWebGLMethodsHelpers.h
@@ -1,30 +1,21 @@
 #pragma once
-// it should be included only in EXWebGLMethods.cpp
+
+#include "pch.h"
 
 #include "EXGLNativeContext.h"
 #include "EXWebGLRenderer.h"
 
-#ifdef __ANDROID__
-#include <GLES3/gl3.h>
-#include <GLES3/gl3ext.h>
-#endif
-#ifdef __APPLE__
-#include <OpenGLES/EAGL.h>
-#include <OpenGLES/ES3/gl.h>
-#include <OpenGLES/ES3/glext.h>
-#endif
-
 namespace expo {
 namespace gl_cpp {
 
-template <typename Func>
+template<typename Func>
 inline jsi::Value exglGetActiveInfo(
-    EXGLContext *ctx,
-    jsi::Runtime &runtime,
-    EXGLObjectId fProgram,
-    GLuint index,
-    GLenum lengthParam,
-    Func glFunc) {
+  EXGLContext *ctx,
+  jsi::Runtime &runtime,
+  EXGLObjectId fProgram,
+  GLuint index,
+  GLenum lengthParam,
+  Func glFunc) {
   if (fProgram == 0) {
     return nullptr;
   }
@@ -49,21 +40,21 @@ inline jsi::Value exglGetActiveInfo(
   }
 
   jsi::Object jsResult =
-      createWebGLObject(runtime, EXWebGLClass::WebGLActiveInfo, {}).asObject(runtime);
+    createWebGLObject(runtime, EXWebGLClass::WebGLActiveInfo, {}).asObject(runtime);
   jsResult.setProperty(runtime, "name", jsi::String::createFromUtf8(runtime, name));
   jsResult.setProperty(runtime, "size", size);
   jsResult.setProperty(runtime, "type", static_cast<double>(type));
   return jsResult;
 }
 
-template <typename Func, typename... T>
+template<typename Func, typename... T>
 inline jsi::Value exglCall(EXGLContext *ctx, Func func, T &&... args) {
   ctx->addToNextBatch([=, args = std::make_tuple(std::forward<T>(args)...)] {
     return std::apply(func, std::move(args));
   });
 }
 
-template <typename Func, typename T>
+template<typename Func, typename T>
 inline jsi::Value
 exglUniformv(EXGLContext *ctx, Func func, GLuint uniform, size_t dim, std::vector<T> &&data) {
   ctx->addToNextBatch([=, data{std::move(data)}] {
@@ -72,21 +63,21 @@ exglUniformv(EXGLContext *ctx, Func func, GLuint uniform, size_t dim, std::vecto
   return nullptr;
 }
 
-template <typename Func, typename T>
+template<typename Func, typename T>
 inline jsi::Value exglUniformMatrixv(
-    EXGLContext *ctx,
-    Func func,
-    GLuint uniform,
-    GLboolean transpose,
-    size_t dim,
-    std::vector<T> &&data) {
+  EXGLContext *ctx,
+  Func func,
+  GLuint uniform,
+  GLboolean transpose,
+  size_t dim,
+  std::vector<T> &&data) {
   ctx->addToNextBatch([=, data{std::move(data)}] {
     func(uniform, static_cast<int>(data.size() / dim), transpose, data.data());
   });
   return nullptr;
 }
 
-template <typename Func, typename T>
+template<typename Func, typename T>
 inline jsi::Value
 exglVertexAttribv(EXGLContext *ctx, Func func, GLuint index, std::vector<T> &&data) {
   ctx->addToNextBatch([=, data{std::move(data)}] { func(index, data.data()); });
@@ -101,10 +92,10 @@ exglIsObject(EXGLContext *ctx, EXGLObjectId id, std::function<GLboolean(GLuint)>
 }
 
 inline jsi::Value exglGenObject(
-    EXGLContext *ctx,
-    jsi::Runtime &runtime,
-    std::function<void(GLsizei, EXGLObjectId *)> func,
-    EXWebGLClass webglClass) {
+  EXGLContext *ctx,
+  jsi::Runtime &runtime,
+  std::function<void(GLsizei, EXGLObjectId *)> func,
+  EXWebGLClass webglClass) {
   auto id = ctx->addFutureToNextBatch(runtime, [=] {
     GLuint buffer;
     func(1, &buffer);
@@ -114,10 +105,10 @@ inline jsi::Value exglGenObject(
 }
 
 inline jsi::Value exglCreateObject(
-    EXGLContext *ctx,
-    jsi::Runtime &runtime,
-    std::function<GLuint()> func,
-    EXWebGLClass webglClass) {
+  EXGLContext *ctx,
+  jsi::Runtime &runtime,
+  std::function<GLuint()> func,
+  EXWebGLClass webglClass) {
   auto id = ctx->addFutureToNextBatch(runtime, [=] { return func(); });
   return createWebGLObject(runtime, webglClass, {std::move(id)});
 }
@@ -129,9 +120,9 @@ exglDeleteObject(EXGLContext *ctx, EXGLObjectId id, std::function<void(EXGLObjec
 }
 
 inline jsi::Value exglDeleteObject(
-    EXGLContext *ctx,
-    EXGLObjectId id,
-    std::function<void(GLsizei, const EXGLObjectId *)> func) {
+  EXGLContext *ctx,
+  EXGLObjectId id,
+  std::function<void(GLsizei, const EXGLObjectId *)> func) {
   ctx->addToNextBatch([=] {
     GLuint buffer = ctx->lookupObject(id);
     func(1, &buffer);
